initial implementation of resfile
added two functions to sampler.py to read the resfile and enforce it's constraints logits.

added a new dictionary to atoms.py converting command codes from resfile to lists of amino acids that need to be minimized.

added an argument to the run_manager.py to include --resfile and the path to the file

added an example resfile.txt