changed the way logits are modified to control the distribution
Previously, sampler.py's enforce_resfile would assign 0 to the logits that had to be avoived with a resfile.

However, pytorch's Categorical distribution normalizes the passed logits, which means that assigning 0 still makes some residues probable.

Numpy's negative infinity seems to be a solution because it doesn't get normalized by the Categorical object when passed as logits.
This way, the Categorical distribution only normalizes the values that are non-negative-infinity (doesn't work the same way with a positive infinity).