added the single residue handling in resfile
now, when the resfile has a command with a single amino acid for a particular residue (ex. 5 PIKAA C or eq. with NOTAA), that restriction can be used to initialize the sequence

NOTE: to initialize the sequence with resfile, you MUST use --randomize 0 when running run.py (to skip running the baseline model)

updated sampler.py to read in single residue restrictions

updated pyrosetta_util.py to take in the resfile restrictions for sequence initialization

updated resfile.py to identify and pass residues with single amino acids

added resfile3.txt to test the functionality on 3mx7_gt.pdb to initialize with only CYS